#3 - Capture Audio from AudioRender, but deprecated
AudioRender is deprecated, because non-init issue can't be fixed
- add getting valid processes

+++ Recording/Video/WGC/WindowEnumerationHelper.cs
@@ -128,7 +128,7 @@ namespace BetterLiveScreen.Recording.Video.WGC
                 return false;
             }
 
-            var style = (WindowStyles)(uint)GetWindowLongPtr(hwnd, (int)GWL.GWL_STYLE).ToInt32();
+            var style = (WindowStyles)(long)GetWindowLongPtr(hwnd, (int)GWL.GWL_STYLE).ToInt64();
             if (style.HasFlag(WindowStyles.WS_DISABLED))
             {
                 return false;
